mutsa-bootcamp-vanilla-project/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
python-jose==3.5.0
passlib==1.7.4
bcrypt<4.0.0
python-multipart==0.0.20
pydantic==2.11.7
supabase